+++ docs/conf.py
@@ -100,6 +100,7 @@ nitpick_ignore = [
     ("py:class", "CarrierT"),
     ("py:obj", "opentelemetry.propagators.textmap.CarrierT"),
     ("py:obj", "Union"),
+    ("py:data", "typing.Union"),
     (
         "py:class",
         "opentelemetry.sdk.metrics._internal.instrument._Synchronous",
@@ -144,8 +145,11 @@ nitpick_ignore = [
         "py:class",
         "opentelemetry.proto.collector.metrics.v1.metrics_service_pb2.ExportMetricsServiceRequest",
     ),
-    ("py:class", "opentelemetry.sdk._logs._internal.export.LogExporter"),
-    ("py:class", "opentelemetry.sdk._logs._internal.export.LogExportResult"),
+    ("py:class", "opentelemetry.sdk._logs._internal.export.LogRecordExporter"),
+    (
+        "py:class",
+        "opentelemetry.sdk._logs._internal.export.LogRecordExportResult",
+    ),
     (
         "py:class",
         "opentelemetry.proto.collector.logs.v1.logs_service_pb2.ExportLogsServiceRequest",